20201119(1)

+++ printf/libft/include/libft.h
@@ -6,7 +6,7 @@
 /*   By: yjung <yjung@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/01 16:20:17 by yjung             #+#    #+#             */
-/*   Updated: 2020/11/11 18:34:14 by yjung            ###   ########.fr       */
+/*   Updated: 2020/11/19 01:09:26 by yjung            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -70,6 +70,6 @@ void				ft_lstiter(t_list *lst, void (*f)(void *));
 t_list				*ft_lstmap(t_list *lst, void *(*f)(void *), \
 		void (*del)(void *));
 unsigned long long	ft_pow(int b);
-unsigned long long	ft_int_cnt(unsigned long long num);
+int					ft_int_cnt(unsigned long long num);
 
 #endif
